requirements.txt
-e git+git@github.com:wolkiewiczk/cookie-cutter.git@b37911fb87a0f428bc65ce54347ef47cf2be0fa9#egg=Cookie_cutter
Flask==1.1.1
Pillow==6.2.1
gunicorn==20.0.0
